Changed how entity prefab instance work
Now it creates a temporary entity with the prefab instance data and then the system processes that and remove the temporary entity. That is to avoid the initial search in all root objects for entity prefab instances and also to allow instantiating entity prefab instances on the run and process those in the same system.